Move config types to live with the rest of them
